refactor(frontend): split active room streaming
